+++ contracts/mocks/VoucherHubV3Mock.sol
@@ -3,17 +3,17 @@
 
 pragma solidity ^0.8.20;
 
-import {VoucherHub} from "../VoucherHub.sol";
+import {VoucherHubV2} from "../VoucherHubV2.sol";
 
 /**
  * @notice This contract is for upgradeability testing purposes only.
  */
 
-contract VoucherHubV2Mock is VoucherHub {
+contract VoucherHubV3Mock is VoucherHubV2 {
     bytes32 private constant VOUCHER_HUB_STORAGE_LOCATION =
         0xfff04942078b704e33df5cf14e409bc5d715ca54e60a675b011b759db89ef800;
 
-    struct VoucherHubStorageV2 {
+    struct VoucherHubStorageV3 {
         address _iexecPoco;
         address _voucherBeacon;
         /// @dev This hash should be updated when `VoucherProxy` is updated.
@@ -24,17 +24,17 @@ contract VoucherHubV2Mock is VoucherHub {
         string _foo;
     }
 
-    function initializeV2(string calldata bar) public reinitializer(2) {
-        VoucherHubStorageV2 storage $ = _getVoucherHubStorageV2();
+    function initializeV3(string calldata bar) public reinitializer(3) {
+        VoucherHubStorageV3 storage $ = _getVoucherHubStorageV3();
         $._foo = bar;
     }
 
     function foo() external view returns (string memory) {
-        VoucherHubStorageV2 storage $ = _getVoucherHubStorageV2();
+        VoucherHubStorageV3 storage $ = _getVoucherHubStorageV3();
         return $._foo;
     }
 
-    function _getVoucherHubStorageV2() private pure returns (VoucherHubStorageV2 storage $) {
+    function _getVoucherHubStorageV3() private pure returns (VoucherHubStorageV3 storage $) {
         assembly ("memory-safe") {
             $.slot := VOUCHER_HUB_STORAGE_LOCATION
         }