+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Changelog
 
 ## vNEXT
-- Add `dealId` to `TaskClaimedWithVoucher` event. (#46)
+- Add `dealId` to `TaskClaimedWithVoucher` event. (#50)
+- Remove references to blockscout v5. (#49)
+
+## v1.0.0
+
+### What's new?
+- Allow users to access resources of the iExec network via a sponsorship voucher.
+
+### More details
 - Upgrade Solidity Compiler to `v0.8.27`. (#45)
 - Bump dependencies: (#44)
     - `@openzeppelin/hardhat-upgrades`, `hardhat`, `ethers`, `prettier`, and others [minor version bump]